requirements.txt
fastapi
uvicorn
requests
pyjwt
google-generativeai
python-dotenv
